Refactor all moment to use date except s3 metadata, and fix bug with modify listing update endpoint
