+++ src/controllers/listingComments.ts
@@ -1,5 +1,5 @@
 import moment from 'moment-timezone';
-import { db } from '../database/db';
+import { db } from '../database';
 import { asyncHandler } from '../middleware';
 import { cleanseData, ErrorResponse, parseSqlUpdateStmt } from '../utils';
 
@@ -117,7 +117,7 @@ export const updateListingComment = asyncHandler(async (req, res, next) => {
 
     const data = {
         comment,
-        updated_on: moment.tz(process.env.DEFAULT_TIMEZONE).format(),
+        updated_on: moment.tz(process.env.DEFAULT_TIMEZONE).toDate(),
     };
 
     cleanseData(data);
@@ -144,7 +144,7 @@ export const deactivateListingComment = asyncHandler(async (req, res, next) => {
     }
 
     const data = {
-        deleted_on: moment.tz(process.env.DEFAULT_TIMEZONE).format(),
+        deleted_on: moment.tz(process.env.DEFAULT_TIMEZONE).toDate(),
     };
 
     const deactivateListingCommentQuery = parseSqlUpdateStmt(data, 'listingcomments', 'WHERE listing_comment_id = $1 RETURNING *', [req.params.id]);
